fix imports, move fsst+dict bench into fsst
