+++ encodings/fsst/Cargo.toml
@@ -20,6 +20,7 @@ workspace = true
 async-trait = { workspace = true }
 fsst-rs = { workspace = true }
 prost = { workspace = true }
+rand = { workspace = true, optional = true }
 vortex-array = { workspace = true }
 vortex-buffer = { workspace = true }
 vortex-dtype = { workspace = true }
@@ -27,6 +28,9 @@ vortex-error = { workspace = true }
 vortex-mask = { workspace = true }
 vortex-scalar = { workspace = true }
 vortex-vector = { workspace = true }
+
+[features]
+test-harness = ["dep:rand", "vortex-array/test-harness"]
 
 [dev-dependencies]
 divan = { workspace = true }
@@ -38,3 +42,8 @@ vortex-array = { workspace = true, features = ["test-harness"] }
 [[bench]]
 name = "fsst_compress"
 harness = false
+
+[[bench]]
+name = "chunked_dict_array_builder"
+harness = false
+required-features = ["test-harness"]
