+++ encodings/fsst/src/lib.rs
@@ -17,6 +17,8 @@ mod compress;
 mod compute;
 mod ops;
 mod serde;
+#[cfg(feature = "test-harness")]
+pub mod test_utils;
 #[cfg(test)]
 mod tests;
