+++ vortex-array/benches/dict_compress.rs
@@ -5,10 +5,10 @@
 
 use divan::Bencher;
 use rand::distr::{Distribution, StandardUniform};
+use vortex_array::arrays::dict_test::{gen_primitive_for_dict, gen_varbin_words};
 use vortex_array::arrays::{VarBinArray, VarBinViewArray};
+use vortex_array::builders::dict::dict_encode;
 use vortex_array::compute::warm_up_vtables;
-use vortex_dict::builders::dict_encode;
-use vortex_dict::test::{gen_primitive_for_dict, gen_varbin_words};
 use vortex_dtype::NativePType;
 
 fn main() {
