+++ vortex-array/benches/dict_mask.rs
@@ -7,9 +7,8 @@ use divan::Bencher;
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
 use vortex_array::IntoArray;
-use vortex_array::arrays::PrimitiveArray;
+use vortex_array::arrays::{DictArray, PrimitiveArray};
 use vortex_array::compute::{mask, warm_up_vtables};
-use vortex_dict::DictArray;
 use vortex_mask::Mask;
 
 fn main() {
