+++ vortex-array/src/arrays/dict/compute/is_constant.rs
@@ -1,13 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-use vortex_array::compute::{
-    IsConstantKernel, IsConstantKernelAdapter, IsConstantOpts, is_constant_opts,
-};
-use vortex_array::register_kernel;
 use vortex_error::VortexResult;
 
-use crate::{DictArray, DictVTable};
+use super::{DictArray, DictVTable};
+use crate::compute::{IsConstantKernel, IsConstantKernelAdapter, IsConstantOpts, is_constant_opts};
+use crate::register_kernel;
 
 impl IsConstantKernel for DictVTable {
     fn is_constant(&self, array: &DictArray, opts: &IsConstantOpts) -> VortexResult<Option<bool>> {
