+++ vortex-btrblocks/Cargo.toml
@@ -25,7 +25,6 @@ vortex-array = { workspace = true }
 vortex-buffer = { workspace = true }
 vortex-datetime-parts = { workspace = true }
 vortex-decimal-byte-parts = { workspace = true }
-vortex-dict = { workspace = true }
 vortex-dtype = { workspace = true }
 vortex-error = { workspace = true }
 vortex-fastlanes = { workspace = true }
