+++ vortex-file/src/lib.rs
@@ -107,11 +107,11 @@ pub use forever_constant::*;
 pub use open::*;
 pub use strategy::*;
 use vortex_alp::{ALPEncoding, ALPRDEncoding};
+use vortex_array::arrays::DictEncoding;
 use vortex_array::{ArraySessionExt, EncodingRef};
 use vortex_bytebool::ByteBoolEncoding;
 use vortex_datetime_parts::DateTimePartsEncoding;
 use vortex_decimal_byte_parts::DecimalBytePartsEncoding;
-use vortex_dict::DictEncoding;
 use vortex_fastlanes::{BitPackedEncoding, DeltaEncoding, FoREncoding, RLEEncoding};
 use vortex_fsst::FSSTEncoding;
 use vortex_pco::PcoEncoding;
