+++ vortex-layout/src/layouts/dict/writer.rs
@@ -10,10 +10,10 @@ use async_trait::async_trait;
 use futures::future::BoxFuture;
 use futures::stream::{BoxStream, once};
 use futures::{FutureExt, Stream, StreamExt, TryStreamExt, pin_mut, try_join};
+use vortex_array::arrays::DictEncoding;
+use vortex_array::builders::dict::{DictConstraints, DictEncoder, dict_encoder};
 use vortex_array::{Array, ArrayContext, ArrayRef};
 use vortex_btrblocks::BtrBlocksCompressor;
-use vortex_dict::DictEncoding;
-use vortex_dict::builders::{DictConstraints, DictEncoder, dict_encoder};
 use vortex_dtype::Nullability::NonNullable;
 use vortex_dtype::{DType, PType};
 use vortex_error::{VortexError, VortexResult, vortex_err};
