+++ vortex/Cargo.toml
@@ -27,7 +27,6 @@ vortex-buffer = { workspace = true }
 vortex-bytebool = { workspace = true }
 vortex-datetime-parts = { workspace = true }
 vortex-decimal-byte-parts = { workspace = true }
-vortex-dict = { workspace = true, features = ["arrow"] }
 vortex-dtype = { workspace = true, default-features = true }
 vortex-error = { workspace = true }
 vortex-fastlanes = { workspace = true }
