+++ vortex/benches/common_encoding_tree_throughput.rs
@@ -9,13 +9,12 @@ use divan::Bencher;
 use divan::counter::BytesCount;
 use mimalloc::MiMalloc;
 use rand::{Rng, SeedableRng};
-use vortex::arrays::{PrimitiveArray, TemporalArray, VarBinArray, VarBinViewArray};
+use vortex::arrays::{DictArray, PrimitiveArray, TemporalArray, VarBinArray, VarBinViewArray};
 use vortex::compute::cast;
 use vortex::dtype::datetime::TimeUnit;
 use vortex::dtype::{DType, PType};
 use vortex::encodings::alp::alp_encode;
 use vortex::encodings::datetime_parts::{DateTimePartsArray, split_temporal};
-use vortex::encodings::dict::DictArray;
 use vortex::encodings::fastlanes::FoRArray;
 use vortex::encodings::fsst::{FSSTArray, fsst_compress, fsst_train_compressor};
 use vortex::encodings::runend::RunEndArray;
