+++ vortex/benches/single_encoding_throughput.rs
@@ -11,10 +11,10 @@ use mimalloc::MiMalloc;
 use rand::prelude::IndexedRandom;
 use rand::{Rng, SeedableRng};
 use vortex::arrays::{PrimitiveArray, VarBinViewArray};
+use vortex::builders::dict::dict_encode;
 use vortex::compute::cast;
 use vortex::dtype::PType;
 use vortex::encodings::alp::{RDEncoder, alp_encode};
-use vortex::encodings::dict::builders::dict_encode;
 use vortex::encodings::fastlanes::{DeltaArray, FoRArray, delta_compress};
 use vortex::encodings::fsst::{fsst_compress, fsst_train_compressor};
 use vortex::encodings::pco::PcoArray;
