+++ vortex/src/lib.rs
@@ -30,10 +30,9 @@ pub mod encodings {
     pub use vortex_zstd as zstd;
     pub use {
         vortex_alp as alp, vortex_bytebool as bytebool, vortex_datetime_parts as datetime_parts,
-        vortex_decimal_byte_parts as decimal_byte_parts, vortex_dict as dict,
-        vortex_fastlanes as fastlanes, vortex_fsst as fsst, vortex_pco as pco,
-        vortex_runend as runend, vortex_sequence as sequence, vortex_sparse as sparse,
-        vortex_zigzag as zigzag,
+        vortex_decimal_byte_parts as decimal_byte_parts, vortex_fastlanes as fastlanes,
+        vortex_fsst as fsst, vortex_pco as pco, vortex_runend as runend,
+        vortex_sequence as sequence, vortex_sparse as sparse, vortex_zigzag as zigzag,
     };
 }
